Replace the legacy filter with a compatibility layer
