+++ ranger/core/filter_stack.py
@@ -14,11 +14,31 @@ except ImportError:
 # pylint: enable=invalid-name
 from os.path import abspath
 
-from ranger.container.directory import accept_file, InodeFilterConstants
 from ranger.core.shared import FileManagerAware
 from ranger.ext.hash import hash_chunks
 
 # pylint: disable=too-few-public-methods
+
+
+class InodeFilterConstants(object):
+    DIRS = "d"
+    FILES = "f"
+    LINKS = "l"
+
+
+def accept_file(fobj, filters):
+    """
+    Returns True if file shall be shown, otherwise False.
+    Parameters:
+        fobj - an instance of FileSystemObject
+        filters - an array of lambdas, each expects a fobj and
+                  returns True if fobj shall be shown,
+                  otherwise False.
+    """
+    for filt in filters:
+        if filt and not filt(fobj):
+            return False
+    return True
 
 
 class BaseFilter(object):
